requirements.txt
alembic==1.13.1
altair==5.3.0
APScheduler==3.10.4
attrs==23.2.0
bcrypt==4.1.3
bidict==0.23.1
blinker==1.8.2
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect2==0.7.0
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
fonttools==4.51.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lxml==5.2.2
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysqlclient==2.2.4
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pdfkit==1.0.0
pillow==10.3.0
pinecone==4.0.0
protobuf==4.25.3
pyarrow==16.0.0
pycparser==2.22
pydeck==0.9.0
pydyf==0.10.0
Pygments==2.18.0
PyJWT==2.8.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.4.2
scipy==1.13.0
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.30
stripe==9.6.0
tenacity==8.3.0
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
watchdog==4.0.0
weasyprint==62.1
webencodings==0.5.1
Werkzeug==3.0.3
wsproto==1.2.0
zopfli==0.2.3
